backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-core==2.23.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.17
websockets==14.1
aiosmtplib==3.0.2
email-validator==2.2.0
python-dotenv==1.0.1